html/semantics/embedded-content/the-video-element/video-loading-lazy-load-when-visible.html

<!DOCTYPE html>
<link rel="author" title="Squarespace" href="http://www.squarespace.com/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
  #target {
    position: absolute;
    top: 300vh;
  }
</style>

<video id="target" src="/media/A4.mp4" loading="lazy" width="100" height="100"></video>

<script>
  const video = document.getElementById("target");

  const t = async_test(
    "Video with loading=lazy does not fetch while not visible, then fetches once scrolled into view"
  );

  const url = new URL(video.src, location.href).href;


  // Step 1: Not visible => should not fetch.
  t.step_timeout(() => {
    t.step(() => {
      assert_equals(performance.getEntriesByName(url).length, 0);
    });

    // Step 2: Scroll into view => should fetch.
    video.scrollIntoView();

    t.step_timeout(() => {
      t.step(() => {
        assert_greater_than(performance.getEntriesByName(url).length, 0);
      });
      t.done();
    }, 1000);
  }, 1000);
</script>
</body>
</html>
